+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog Fab-manager
 
 - Fix a bug: during the upgrade, unable to detect the target when not explicitly specified
+- Fix a bug: unable to sync users on stripe if one of them have an invalid email address
 
 ## v5.0.4 2021 June 15